requirements.txt
streamlit>=1.35.0
pandas
gspread
google-auth
googlemaps
openai
python-dotenv
